pyproject.toml
[project]
name = "senac-ia-uc15-nlp-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Pablo Henrique",email = "35077005+pablohls@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "datasets (>=4.8.4,<5.0.0)",
    "chromadb (>=1.2.0,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "langchain-core (>=0.3.79,<0.4.0)",
    "pypdf (>=6.1.2,<7.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "transformers[torch] (>=5.4.0,<6.0.0)",
    "peft (>=0.18.1,<0.19.0)",
    "whisperx (>=3.3.4)",
    "pyannote.audio (>=3.3.2,<4.0.0)",
    "fpdf2 (>=2.8.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)"
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.2.0"

